Move delegation state maintenance to stake.rs
